LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/pseudo-get-computed-style-expected.txt

CONSOLE MESSAGE: Error: assert_equals: container(target) expected "multiply" but got "normal"
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error

Harness Error (FAIL), message = Unhandled rejection: Type error

TIMEOUT position property of pseudo elements Test timed out
NOTRUN properties of pseudo elements outside of transition
